requirements.txt
Django==5.2.5
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
django-filter==23.5
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
coverage==7.6.3
pytest==8.3.3
pytest-django==4.9.0
factory-boy==3.3.1 
channels==4.0.0
websocket-client==1.6.4
python-dotenv==1.0.0
channels-redis==4.1.0
requests==2.31.0
python-dotenv==1.0.0